Move Promise support to directives.
Removes built-in Promise support. Adds an async() directive as a successor to until() that is requires to render Promises. async() handles race conditions, works with attribute, boolean attribute, property and event bindings. It also works as a nested directive value with other directives like repeat().

While this only removes a small amount of current core library code, more core code would have be required to handle attributes, etc., and the race conditions, mainly because current Part factoring has us sharing very little code between Part implementations. It’s possible that we could refactor Parts in the future to share more value-handling code and bring back built-in Promise support in the future. Adding that would not be a breaking change, but we need to remove support before 1.0 since removing later would be a breaking change.

This change keeps directives/until.js to re-export async to ease the transition. It’s deprecated and will be removed before 1.0.